fix(push): fourth-round claude-review feedback for DEQ-283
- Actually USE the `DeviceTokenRegistrationBody` Encodable struct in
  `postDeviceToken` (prior edit defined but never wired). JSON body
  encoded via `JSONEncoder`; `[String: Any]` + `JSONSerialization`
  path is gone.
- Drop redundant `if let authToken` after exhaustive `do-catch` in
  both `postDeviceToken` and `deregisterOnSignOut`. The binding is
  now `String` (not `String?`).
- `SilentPushPayload.sentAtMs` is now `Int64?` (was `Double?`) to
  match the CLAUDE.md project-wide ms-as-Int64 convention. AppDelegate
  extracts via `NSNumber.int64Value` so APNs serialisation quirks
  (NSNumber vs Int vs Double) never lose precision.
- `SilentSyncOutcome` is now genuinely `private`; doc comment no
  longer references a non-existent test that touches it.
- `AppContext` extracted into its own file `Services/AppContext.swift`.
  General-purpose service locator referenced from multiple files
  shouldn't live inside `PushNotificationService.swift`.

Tests / verification:
- 17 PushNotificationServiceTests cases still green.
- SwiftLint 0 violations.
- iOS + macOS builds succeed.

Refs DEQ-276
Fixes DEQ-283
